move interfaces to separate pkgs
